Rename: 회원 도메인 이름 변경 및 패키지 이동
domain 패키지 안에 entity 패키지로 이동 시켰다.
