+++ src/main/java/com/devwinter/userservice/domain/entity/User.java
@@ -1,11 +1,13 @@
-package com.devwinter.userservice.domain;
+package com.devwinter.userservice.domain.entity;
 
 import lombok.AccessLevel;
 import lombok.Builder;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
+@Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
 @Table(
