+++ src/main/java/com/devwinter/userservice/domain/repository/UserCommandRepository.java
@@ -1,0 +1,9 @@
+package com.devwinter.userservice.domain.repository;
+
+import com.devwinter.userservice.domain.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserCommandRepository extends JpaRepository<User, Long> {
+
+    boolean existsByEmail(String email);
+}
