--- src/main/java/com/devwinter/userservice/domain/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package com.devwinter.userservice.domain.repository;
-
-import com.devwinter.userservice.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-}
